+++ cca_zoo/models/rcca.py
@@ -111,3 +111,21 @@ class CCA(rCCA):
         :param latent_dims: number of latent dimensions to learn
         """
         super().__init__(latent_dims=latent_dims, c=[0.0, 0.0])
+
+
+def _pca_data(*views: Tuple[np.ndarray, ...]):
+    """
+    Since most methods require zero-mean data, demean_data() is used to demean training data as well as to apply this
+    demeaning transformation to out of sample data
+
+    :param views: numpy arrays with the same number of rows (samples) separated by commas
+    """
+    views_U = []
+    views_S = []
+    views_Vt = []
+    for i, view in enumerate(views):
+        U, S, Vt = np.linalg.svd(view, full_matrices=False)
+        views_U.append(U)
+        views_S.append(S)
+        views_Vt.append(Vt)
+    return views_U, views_S, views_Vt
